Added PlayerTempIdentificationPacket
This wouldn't be necessary if processing.net didn't suck massive dick,
but it does. The client has no idea what its IP is, unfortunately, so
without a way for the server to be able to send packets directly to
clients, we are instead forced to come up with a makeshift way to ID
each packet; and this is the solution, for now at least.